pyproject.toml
[tool.poetry]
name = "arenabackend"
version = "0.1.0"
description = "Arena backend"
authors = ["yassinekdi"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
